ex3_opengl_car_viewer/requirements.txt
PyOpenGL
PyOpenGL_accelerate
glfw
numpy